#226 standardize arcade examples on GameApp module layout FIXED
