+++ examples/flappy_bird/Cargo.toml
@@ -10,7 +10,7 @@ description = "Flappy Bird on-chain game example using cougr-core ECS framework"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-cougr-core = "1.0.0"
+cougr-core = "1.1.0"
 soroban-sdk = "25.1.0"
 
 [dev-dependencies]
