+++ examples/geometry_dash/Cargo.toml
@@ -10,7 +10,7 @@ description = "Geometry Dash on-chain game example using cougr-core ECS framewor
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-cougr-core = "1.0.0"
+cougr-core = "1.1.0"
 soroban-sdk = "25.1.0"
 
 [dev-dependencies]
