+++ examples/geometry_dash/README.md
@@ -1,48 +1,67 @@
-# Geometry Dash Soroban Contract
+# Geometry Dash On-Chain Game
 
-A rhythm platformer game implemented as a Soroban smart contract using the `Cougr` ECS framework.
+> **Transitional example**: This example uses an older Cougr pattern and is preserved
+> for compatibility reference. For the current recommended approach, see `snake`.
 
-## Features
+## Purpose and pattern
 
-- **ECS Architecture**: Uses `Cougr` for clean separation of components and systems.
-- **Tick-based Simulation**: Deterministic game logic that advances in discrete steps.
-- **Multiple Player Modes**: Support for Cube, Ship, Wave, and Ball modes, each with unique physics.
-- **Collision System**: Detects collisions with spikes, blocks, and mode-switching portals.
-- **On-chain State**: Game world and player progress are persisted in Soroban storage.
+This example demonstrates a rhythm-platformer fixed-tick runner on Soroban with Cougr ECS concepts. It remains transitional while the arcade examples converge on the canonical `snake` `GameApp` architecture.
 
-## Game Physics
+## Public contract API
 
-The game operates on a fixed-fixed point coordinate system (scaled by 1000).
+| Function | Parameters | Return type | Description |
+|---|---|---:|---|
+| `init_game` | `player: Address, level_id: u32` | `()` | Initializes a player level. |
+| `jump` | `player: Address` | `()` | Applies the current mode input. |
+| `update_tick` | `player: Address` | `()` | Advances the level simulation by one tick. |
+| `get_state` | `player: Address` | `GameStatus` | Returns Playing, Crashed, or Completed. |
+| `get_pos` | `player: Address` | `(i32, i32)` | Returns player position. |
+| `get_score` | `player: Address` | `u32` | Returns distance-based score. |
+| `get_mode` | `player: Address` | `u32` | Returns current mode identifier. |
 
-- **Cube**: Gravity pulls down, tap to jump.
-- **Ship**: Constant gravity, hold (multi-tap) to fly up.
-- **Wave**: Oscillates at 45 degrees up when held, 45 degrees down when released.
-- **Ball**: Tapping switches gravity direction.
+## Architecture overview
 
-## API
-
-- `init_game(player: Address, level_id: u32)`: Initializes the level and player state.
-- `jump(player: Address)`: Triggers the "action" input for the current player mode.
-- `update_tick(player: Address)`: Advances the game state by one frame.
-- `get_state(player: Address) -> GameStatus`: Returns `Playing`, `Crashed`, or `Completed`.
-- `get_score(player: Address) -> u32`: Returns current distance-based score.
-- `get_mode(player: Address) -> u32`: Returns current player mode identifier.
-
-## Testing
-
-Run unit tests:
-```bash
-cargo test
+```text
+contract entrypoint
+  ├─ reads game state from Soroban storage
+  ├─ applies input or tick systems
+  └─ writes updated state back to storage
 ```
 
-Build the contract:
+Player physics, mode, obstacle, and status components live in `components.rs`; motion/collision behavior lives in `systems.rs`.
+
+## Storage model
+
+| Storage class | Data | Why |
+|---|---|---|
+| Instance storage | Per-contract game state where used by this example. | Keeps small arcade state close to the contract instance. |
+| Persistent storage | Player- or world-scoped state where the example needs durable keyed state. | Keeps game progress available across invocations. |
+| Temporary storage | Not used. | The examples favor deterministic recalculation over ephemeral caches. |
+
+## Main gameplay flow
+
+1. Call the initialization function to create the starting state.
+2. Submit an input action such as movement, jump, flap, rotation, or shoot.
+3. Call the tick/update function to run deterministic simulation logic.
+4. Query public getters for score, position, active state, or terminal status.
+5. Stop when the game-over/completed condition is reached, or reset/reinitialize where supported.
+
+## Cougr APIs used
+
+- `ComponentTrait` and custom component modules document the ECS data boundary.
+- `SimpleWorld`, `SimpleQueryBuilder`, `GameApp`, `ScheduleStage`, or `SystemConfig` are used where this transitional example has already adopted the maintained runtime shape.
+- Auth, privacy, ZK, and standards APIs are intentionally not used; these arcade examples focus on deterministic game logic.
+
+## Build and test commands
+
 ```bash
+cargo test
 stellar contract build
 ```
 
-## Level Structure
+## Known limitations
 
-Obstacles consist of:
-- **Spikes**: Trigger `GameStatus::Crashed` on collision.
-- **Blocks**: Standard solid obstacles.
-- **Portals**: Change the player's `PlayerMode`.
+- Transitional code may preserve older storage or scheduling patterns for compatibility reference.
+- No authentication, matchmaking, real-time rendering, or production randomness is included.
+- One contract instance generally represents one game or one keyed set of player games.
+- For new work, prefer the canonical `snake` module split and `GameApp` tick wiring.
