+++ examples/tetris/src/lib.rs
@@ -1,72 +1,11 @@
 #![no_std]
 
-use soroban_sdk::{contract, contractimpl, contracttype, symbol_short, Env, Vec};
+mod components;
+mod systems;
 
-// We aliasing cougr_core types to avoid confusion if we had local duplicates,
-// but here we just import them.
-// Note: In a real scenario, we'd ensure cougr_core is compatible with soroban-sdk v21.
-use cougr_core::SimpleWorld;
+pub use components::{GameState, Piece, TetrominoShape};
+use soroban_sdk::{contract, contractimpl, symbol_short, Env, Vec};
 
-// --------------------------------------------------------------------------------
-// Data Structures
-// --------------------------------------------------------------------------------
-
-#[contracttype]
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum TetrominoShape {
-    I = 0,
-    J = 1,
-    L = 2,
-    O = 3,
-    S = 4,
-    T = 5,
-    Z = 6,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct Piece {
-    pub shape: TetrominoShape,
-    pub x: i32,
-    pub y: i32,
-    pub rotation: u32, // 0, 1, 2, 3
-}
-
-#[contracttype]
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct GameState {
-    // Board is 20x10. We can represent it as a Vec<Vec<bool>> or flattened.
-    // For Soroban efficiency, maybe Vec<u32> where each u32 is a row?
-    // 20 rows. 10 bits used per row.
-    pub board: Vec<u32>,
-    pub current_piece: Piece,
-    pub next_piece: Piece,
-    pub score: u32,
-    pub level: u32,
-    pub lines_cleared: u32,
-    pub game_over: bool,
-}
-
-// --------------------------------------------------------------------------------
-// ECS Components
-// --------------------------------------------------------------------------------
-
-// We use cougr-core Components to represent the active piece during logic updates.
-// We need to implement serialization for custom components if we want to store them,
-// but for this example, we might use standard types or transient World.
-
-// However, cougr-core v0.0.1 likely requires components to handle Bytes.
-// Let's define a helper to map our Piece to ECS components.
-
-// Position is often a standard component.
-// We'll define a custom component for Tetromino info.
-
-// --------------------------------------------------------------------------------
-// Contract
-// --------------------------------------------------------------------------------
-
-const BOARD_WIDTH: i32 = 10;
-const BOARD_HEIGHT: i32 = 20;
 
 #[contract]
 pub struct TetrisContract;
@@ -78,8 +17,8 @@ impl TetrisContract {
         let board = Vec::from_array(&env, [0u32; 20]); // 20 empty rows
 
         // Spawn initial pieces
-        let current_piece = generate_piece(&env);
-        let next_piece = generate_piece(&env);
+        let current_piece = systems::generate_piece(&env);
+        let next_piece = systems::generate_piece(&env);
 
         let state = GameState {
             board,
@@ -91,7 +30,7 @@ impl TetrisContract {
             game_over: false,
         };
 
-        save_state(&env, &state);
+        systems::save_state(&env, &state);
         state
     }
 
@@ -102,8 +41,8 @@ impl TetrisContract {
             return false;
         }
 
-        if try_move(&env, &mut state, -1, 0, 0) {
-            save_state(&env, &state);
+        if systems::try_move(&env, &mut state, -1, 0, 0) {
+            systems::save_state(&env, &state);
             true
         } else {
             false
@@ -117,8 +56,8 @@ impl TetrisContract {
             return false;
         }
 
-        if try_move(&env, &mut state, 1, 0, 0) {
-            save_state(&env, &state);
+        if systems::try_move(&env, &mut state, 1, 0, 0) {
+            systems::save_state(&env, &state);
             true
         } else {
             false
@@ -132,13 +71,13 @@ impl TetrisContract {
             return false;
         }
 
-        if try_move(&env, &mut state, 0, 1, 0) {
-            save_state(&env, &state);
+        if systems::try_move(&env, &mut state, 0, 1, 0) {
+            systems::save_state(&env, &state);
             true
         } else {
             // Lock piece if it can't move down
-            lock_piece(&env, &mut state);
-            save_state(&env, &state);
+            systems::lock_piece(&env, &mut state);
+            systems::save_state(&env, &state);
             false
         }
     }
@@ -151,8 +90,8 @@ impl TetrisContract {
         }
 
         // Rotation is +1 to index (clockwise)
-        if try_move(&env, &mut state, 0, 0, 1) {
-            save_state(&env, &state);
+        if systems::try_move(&env, &mut state, 0, 0, 1) {
+            systems::save_state(&env, &state);
             true
         } else {
             false
@@ -167,12 +106,12 @@ impl TetrisContract {
         }
 
         let mut dropped = 0;
-        while try_move(&env, &mut state, 0, 1, 0) {
+        while systems::try_move(&env, &mut state, 0, 1, 0) {
             dropped += 1;
         }
 
-        lock_piece(&env, &mut state);
-        save_state(&env, &state);
+        systems::lock_piece(&env, &mut state);
+        systems::save_state(&env, &state);
         dropped
     }
 
@@ -184,11 +123,11 @@ impl TetrisContract {
         }
 
         // Try to move down
-        if !try_move(&env, &mut state, 0, 1, 0) {
-            lock_piece(&env, &mut state);
+        if !systems::try_move(&env, &mut state, 0, 1, 0) {
+            systems::lock_piece(&env, &mut state);
         }
 
-        save_state(&env, &state);
+        systems::save_state(&env, &state);
         state
     }
 
@@ -198,236 +137,6 @@ impl TetrisContract {
             .instance()
             .get(&symbol_short!("game"))
             .expect("Game not initialized")
-    }
-}
-
-// --------------------------------------------------------------------------------
-// Logic & Helpers
-// --------------------------------------------------------------------------------
-
-fn save_state(env: &Env, state: &GameState) {
-    env.storage().instance().set(&symbol_short!("game"), state);
-}
-
-fn generate_piece(env: &Env) -> Piece {
-    // Random shape (0-6)
-    let shape_idx = env.prng().gen_range(0..7);
-    let shape = match shape_idx {
-        0 => TetrominoShape::I,
-        1 => TetrominoShape::J,
-        2 => TetrominoShape::L,
-        3 => TetrominoShape::O,
-        4 => TetrominoShape::S,
-        5 => TetrominoShape::T,
-        _ => TetrominoShape::Z,
-    };
-
-    Piece {
-        shape,
-        x: 3, // Start in middle roughly
-        y: 0,
-        rotation: 0,
-    }
-}
-
-// ECS Integration:
-// We use a ephemeral World to calculate the move validity.
-// This demonstrates usage of cougr-core even if we store state in a simplified struct.
-fn try_move(env: &Env, state: &mut GameState, dx: i32, dy: i32, d_rot: i32) -> bool {
-    // 1. Create ECS World
-    let _world = SimpleWorld::new(env);
-
-    // 2. Define Components
-    // In a full game, we'd have these registered.
-    // Here we map our `Piece` to `Position` and `Shape` (conceptually).
-
-    // Calculate new parameters
-    let new_x = state.current_piece.x + dx;
-    let new_y = state.current_piece.y + dy;
-    let new_rot = (state.current_piece.rotation as i32 + d_rot).rem_euclid(4) as u32;
-
-    // 3. Collision System logic
-    if check_collision(
-        env,
-        &state.board,
-        state.current_piece.shape,
-        new_x,
-        new_y,
-        new_rot,
-    ) {
-        return false;
-    }
-
-    // 4. Update Entity (State)
-    state.current_piece.x = new_x;
-    state.current_piece.y = new_y;
-    state.current_piece.rotation = new_rot;
-
-    true
-}
-
-fn check_collision(
-    _env: &Env,
-    board: &Vec<u32>,
-    shape: TetrominoShape,
-    x: i32,
-    y: i32,
-    rot: u32,
-) -> bool {
-    let coords = get_piece_coords(shape, rot);
-
-    for (cx, cy) in coords {
-        let abs_x = x + cx;
-        let abs_y = y + cy;
-
-        // Wall collision
-        if !(0..BOARD_WIDTH).contains(&abs_x) || abs_y >= BOARD_HEIGHT {
-            return true;
-        }
-
-        // Floor/Existing piece collision
-        if abs_y >= 0 {
-            let row = board.get(abs_y as u32).unwrap_or(0);
-            if (row >> abs_x) & 1 == 1 {
-                return true;
-            }
-        }
-    }
-    false
-}
-
-fn lock_piece(env: &Env, state: &mut GameState) {
-    let coords = get_piece_coords(state.current_piece.shape, state.current_piece.rotation);
-
-    // check game over
-    // If piece is locked and any part is above y=0 (or valid board area start), it's game over?
-    // Actually typically if we can't spawn.
-    // If we lock at y=0, it might be game over.
-
-    let mut game_over = false;
-
-    // Place piece on board
-    for (cx, cy) in coords {
-        let abs_x = state.current_piece.x + cx;
-        let abs_y = state.current_piece.y + cy;
-
-        if abs_y < 0 {
-            game_over = true;
-        } else if abs_y < BOARD_HEIGHT {
-            let mut row = state.board.get(abs_y as u32).unwrap_or(0);
-            row |= 1 << abs_x;
-            state.board.set(abs_y as u32, row);
-        }
-    }
-
-    if game_over {
-        state.game_over = true;
-        return;
-    }
-
-    // Clear lines
-    let mut lines = 0;
-    let mut new_board = Vec::new(env);
-
-    // We rebuild board skipping full lines
-    for i in 0..state.board.len() {
-        let row = state.board.get(i).unwrap();
-        // 10 bits set = 1023 (2^10 - 1)
-        if row == 1023 {
-            lines += 1;
-        } else {
-            new_board.push_back(row);
-        }
-    }
-
-    // Add empty lines at top
-    for _ in 0..lines {
-        new_board.push_front(0); // This might be push_front? Soroban Vec is generic.
-                                 // Actually Soroban Vec `push_front` exists.
-    }
-    state.board = new_board;
-
-    // Score
-    if lines > 0 {
-        let points = match lines {
-            1 => 100,
-            2 => 300,
-            3 => 500,
-            4 => 800,
-            _ => 0,
-        };
-        state.score += points * (state.level + 1);
-        state.lines_cleared += lines;
-        if state.lines_cleared >= state.level * 10 {
-            state.level += 1;
-        }
-    }
-
-    // Spawn new
-    state.current_piece = state.next_piece.clone();
-    state.next_piece = generate_piece(env);
-
-    // Initial collision check for new piece
-    if check_collision(
-        env,
-        &state.board,
-        state.current_piece.shape,
-        state.current_piece.x,
-        state.current_piece.y,
-        state.current_piece.rotation,
-    ) {
-        state.game_over = true;
-    }
-}
-
-// Coordinate definitions for shapes
-// (x, y) offsets relative to pivot
-fn get_piece_coords(shape: TetrominoShape, rot: u32) -> [(i32, i32); 4] {
-    // Simplified rotation system (SRS concepts or basic)
-    // I, J, L, O, S, T, Z
-    match shape {
-        TetrominoShape::I => match rot {
-            0 => [(-1, 0), (0, 0), (1, 0), (2, 0)],
-            1 => [(1, -1), (1, 0), (1, 1), (1, 2)],
-            2 => [(-1, 1), (0, 1), (1, 1), (2, 1)],
-            _ => [(0, -1), (0, 0), (0, 1), (0, 2)],
-        },
-        TetrominoShape::O => [(0, 0), (1, 0), (0, 1), (1, 1)], // No rotation change visually
-        TetrominoShape::T => match rot {
-            0 => [(-1, 0), (0, 0), (1, 0), (0, 1)],
-            1 => [(0, -1), (0, 0), (0, 1), (-1, 0)],
-            2 => [(-1, 0), (0, 0), (1, 0), (0, -1)],
-            _ => [(0, -1), (0, 0), (0, 1), (1, 0)],
-        },
-        // Implement others similarly...
-        // For brevity in this example, mapping placeholders for J, L, S, Z
-        // Using T shape for others to ensure compile, but in real generic implementation we'd fill all.
-        // User asked for "Piece rotation using rotation matrices" or similar.
-        // I will implement all to satisfy "COMPLETE TETRIS GAME LOGIC".
-        TetrominoShape::J => match rot {
-            0 => [(-1, 0), (0, 0), (1, 0), (1, 1)],
-            1 => [(0, -1), (0, 0), (0, 1), (-1, 1)],
-            2 => [(-1, -1), (-1, 0), (0, 0), (1, 0)],
-            _ => [(1, -1), (0, 0), (0, -1), (0, 1)],
-        },
-        TetrominoShape::L => match rot {
-            0 => [(-1, 0), (0, 0), (1, 0), (-1, 1)],
-            1 => [(0, -1), (0, 0), (0, 1), (1, 1)],
-            2 => [(1, -1), (-1, 0), (0, 0), (1, 0)],
-            _ => [(-1, -1), (0, -1), (0, 0), (0, 1)],
-        },
-        TetrominoShape::S => match rot {
-            0 => [(0, 0), (1, 0), (-1, 1), (0, 1)],
-            1 => [(0, -1), (0, 0), (1, 0), (1, 1)],
-            2 => [(0, 0), (1, 0), (-1, 1), (0, 1)], // S/Z 2 states
-            _ => [(0, -1), (0, 0), (1, 0), (1, 1)],
-        },
-        TetrominoShape::Z => match rot {
-            0 => [(-1, 0), (0, 0), (0, 1), (1, 1)],
-            1 => [(1, -1), (1, 0), (0, 0), (0, 1)],
-            2 => [(-1, 0), (0, 0), (0, 1), (1, 1)],
-            _ => [(1, -1), (1, 0), (0, 0), (0, 1)],
-        },
     }
 }
 
@@ -512,4 +221,22 @@ mod test {
 
         assert!(!client.get_state().game_over);
     }
+    #[test]
+    fn test_invalid_action_at_left_wall_returns_false() {
+        let env = Env::default();
+        let client = TetrisContractClient::new(&env, &env.register(TetrisContract, ()));
+        client.init_game();
+        let mut last = true;
+        for _ in 0..16 {
+            last = client.move_left();
+        }
+        assert!(!last);
+    }
+
+    #[test]
+    fn test_gameapp_tick_integration() {
+        let env = Env::default();
+        super::systems::run_gameapp_tick(&env);
+    }
+
 }
